SearchInfoInNewsProject/requirements.txt
langchain
langchain-openai
python-dotenv
streamlit
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
python-magic-bin
OpenAI
